pyproject.toml
[project]
name = "prisoners"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "prisoners", from = "src"}]
requires-python = ">=3.11,<3.14"
dependencies = [
    "a2a-sdk[http-server] (>=0.3.22,<0.4.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "tomli (>=2.3.0,<3.0.0)",
    "google-adk[a2a] (>=1.21.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.1"
